Move the majority of the things in conftest.py into plugins
That way they don't need to be copied and pasted into each test suite
