+++ tests/dev_e2e/pytest.ini
@@ -2,7 +2,4 @@
 filterwarnings =
     ignore:Class property max_ttl is deprecated.*:couchbase.logic.supportability.CouchbaseDeprecationWarning
 
-markers =
-    cbse(num): marks tests to verify a specific CBSE ticket
-
 asyncio_default_fixture_loop_scope = session
