+++ tests/dev_e2e/test_basic_replication.py
@@ -23,6 +23,9 @@ from cbltest.api.test_functions import compare_local_and_remote
 from cbltest.utils import assert_not_null
 
 
+@pytest.mark.min_test_servers(1)
+@pytest.mark.min_sync_gateways(1)
+@pytest.mark.min_couchbase_servers(1)
 class TestBasicReplication(CBLTestClass):
     @pytest.mark.asyncio(loop_scope="session")
     async def test_replicate_non_existing_sg_collections(
