+++ tests/dev_e2e/test_custom_conflict.py
@@ -19,6 +19,9 @@ from cbltest.api.syncgateway import DocumentUpdateEntry
 from cbltest.api.test_functions import compare_local_and_remote
 
 
+@pytest.mark.min_test_servers(1)
+@pytest.mark.min_sync_gateways(1)
+@pytest.mark.min_couchbase_servers(1)
 class TestCustomConflict(CBLTestClass):
     async def do_custom_conflict_test(
         self,
