+++ tests/dev_e2e/test_query_consistency.py
@@ -18,6 +18,9 @@ from cbltest.api.replicator_types import (
 from cbltest.jsonhelper import json_equivalent
 
 
+@pytest.mark.min_test_servers(1)
+@pytest.mark.min_sync_gateways(1)
+@pytest.mark.min_couchbase_servers(1)
 class TestQueryConsistency(CBLTestClass):
     __database: Optional[Database] = None
